requirement.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-admin-autocomplete-filter==0.7.1
django-autoslug==1.9.8
django-countries==7.2.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.6.1
django-nested-inline==0.4.4
django-sql-explorer==2.3
django-tinymce==3.4.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pip-check==2.7
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
terminaltables==3.1.0
xlrd==2.0.1
xlwt==1.3.0
